% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_dat}
\alias{cfa_dat}
\title{A six-variable dataset with 100 cases}
\format{
An object of class \code{data.frame} with 100 rows and 6 columns.
}
\usage{
cfa_dat
}
\description{
A six-variable dataset with 100 cases.
}
\details{
Generated from this model:\preformatted{`f1 =~  .7*x1 + .6*x2 + .8*x3 + .3*x5`
`f2 =~  .2*x1 + .6*x4 + .8*x5 + .7*x6`
`f1 ~~ .2*f2`
}
}
\keyword{datasets}
