% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_dat2}
\alias{cfa_dat2}
\title{Sample data set for a CFA model with an influential case}
\format{
An object of class \code{data.frame} with 100 rows and 7 columns.
}
\usage{
cfa_dat2
}
\description{
A six-variable dataset with 100 cases and potential influential cases
}
\details{
A Sample Model:

\if{html}{\out{<div class="sourceCode">}}\preformatted{`f1 =~  x1 + x2 + x3`
`f2 =~  x4 + x5 + x6`
`f1 ~~ f2`
}\if{html}{\out{</div>}}

Has one or more influential case
}
\keyword{datasets}
