% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_dat_heywood}
\alias{cfa_dat_heywood}
\title{Sample data set for a structural equation model with a negative variance}
\format{
An object of class \code{data.frame} with 60 rows and 6 columns.
}
\usage{
cfa_dat_heywood
}
\description{
A six-variable dataset with 60 cases that has a case that, if not removed,
will result in a negative variance in the solution
}
\details{
A Sample Model:\preformatted{`f1 =~  x1 + x2 + x3`
`f2 =~  x4 + x5 + x6`
`f1 ~~ f2`
}
}
\keyword{datasets}
