% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_rerun}
\alias{cfa_rerun}
\title{The \emph{n} \code{\link[lavaan:sem]{lavaan::sem()}} results on \link{cfa_dat}}
\format{
An object of class \code{list} of length 3.
}
\usage{
cfa_rerun
}
\description{
The \emph{n} \code{\link[lavaan:sem]{lavaan::sem()}} results on \link{cfa_dat}.
}
\details{
The \code{\link[=lavaan_rerun]{lavaan_rerun()}} output fitting the following model to \link{cfa_dat}\preformatted{`f1 =~ x1 + x2 + x3`
`f2 =~ x4 + x5 + x6`
}
}
\keyword{datasets}
