% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_plot.R
\name{gcd_plot}
\alias{gcd_plot}
\alias{md_plot}
\alias{gcd_gof_plot}
\alias{gcd_gof_md_plot}
\title{Plot selected influence statistics and measures of extremeness}
\usage{
gcd_plot(influence_out, cutoff_gcd = 1)

md_plot(influence_out, cutoff_md = NULL, cutoff_md_qchisq = 0.975)

gcd_gof_plot(influence_out, fit_measure, cutoff_gcd = 1, cutoff_fit_measure)

gcd_gof_md_plot(
  influence_out,
  fit_measure,
  cutoff_md = NULL,
  cutoff_fit_measure,
  circle_size = 2,
  cutoff_md_qchisq = 0.975,
  cutoff_gcd = 1
)
}
\arguments{
\item{influence_out}{The output from \code{\link[=influence_stat]{influence_stat()}}.}

\item{cutoff_gcd}{Cases with generalized Cook's distance larger than this
value will be labeled. Default is 1.}

\item{cutoff_md}{Cases with Mahalanobis distance larger than this value
will be labeled. Default is \code{NULL}. If it is \code{NULL},
the (\code{cutoff_md_qchisq} x 100)th percentile of the chi-square distribution with
the degrees of freedom equal to the number of variables
will be used.}

\item{cutoff_md_qchisq}{This value multiplied by 100 is the percentile to be used
for labeling case based on Mahalanobis distance. Default is .975.}

\item{fit_measure}{Specify the fit measure to be used in a plot. Use the name
in the \code{\link[lavaan:fitMeasures]{lavaan::fitMeasures()}} function. No default value.}

\item{cutoff_fit_measure}{Cases with \code{fit_measure} larger than this cutoff in magnitude
will be labeled. No default value and must be specified.}

\item{circle_size}{The size of the largest circle when the size of a circle is controlled
by a statistic.}
}
\value{
A \link{ggplot2} plot. It will not be plot. To plot it, use \code{\link[=plot]{plot()}} on the output.
}
\description{
Get a \code{\link[=influence_stat]{influence_stat()}} output and plot selected statistics
}
\details{
The output of \code{\link[=influence_stat]{influence_stat()}} is simply a matrix. Therefore, this
function will work for any matrix provided.
Row number will be used on the x-axis if applicable. However, case identification
values in the output from \code{\link[=influence_stat]{influence_stat()}} will be used for labeling
individual cases.
}
\examples{
# To be prepared.

}
\seealso{
\code{\link[=influence_stat]{influence_stat()}}.
}
