% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_scores.R
\name{implied_scores}
\alias{implied_scores}
\title{Implied scores for observed outcome variables}
\usage{
implied_scores(fit)
}
\arguments{
\item{fit}{The output from \code{\link[lavaan:lavaan]{lavaan::lavaan()}}, such as \code{\link[lavaan:cfa]{lavaan::cfa()}} and
\code{\link[lavaan:sem]{lavaan::sem()}}.}
}
\value{
A matrix of the implied scores.
}
\description{
Get a \code{\link[lavaan:lavaan]{lavaan::lavaan()}} output and compute the implied scores.
}
\details{
The implied scores for each observed outcome variable are simply computed
in the same way the predicted scores in a linear regression model are
computed.

It currently supports only single-sample path analytic models with only
observed variables.
}
\examples{
library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)

# Compute the implied scores for `m1` and `dv`
fit_implied_scores <- implied_scores(fit)
head(fit_implied_scores)


}
\author{
S. F. Cheung (shufai.cheung@gmail.com)
}
