% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_rerun_check.R
\name{lavaan_rerun_check}
\alias{lavaan_rerun_check}
\title{Check whether the model being fitted is supported}
\usage{
lavaan_rerun_check(fit, print_messages = TRUE)
}
\arguments{
\item{fit}{The output from \code{lavaan}, such as \code{\link[lavaan:cfa]{lavaan::cfa()}} and
\code{\link[lavaan:sem]{lavaan::sem()}}.}

\item{print_messages}{Logical. If \code{TRUE}, will print messages about the
check. If \code{FALSE}, the messages will be attached to the return value
as an attribute. Default is \code{TRUE}.}
}
\value{
A single-element vector.
If confirmed to be supported, will return 0.
If not confirmed be support but may still work, return 1.
If confirmed to be not yet uppported, will return a negative number, the
value of this number without the negative sign is the number of tests failed.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and check whether it is supported.
}
\details{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and check whether it is supported.
}
\examples{

# To be prepared.

}
