% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_exo.R
\name{mahalanobis_exo}
\alias{mahalanobis_exo}
\title{Compute the Mahalanobis distance for each case using only the exogenous
observed variables}
\usage{
mahalanobis_exo(fit)
}
\arguments{
\item{fit}{The output from \code{lavaan}, such as \code{\link[lavaan:cfa]{lavaan::cfa()}} and
\code{\link[lavaan:sem]{lavaan::sem()}}.}
}
\value{
A one-column matrix (a column vector) of the Mahalanobis distance for each
case. The number of rows equal to the data stored in the fit object.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and compute the Mahalanobis distance for each
case using only the exogenous observed variables
}
\details{
For each case, compute the Mahalanobis distance of each case.

Currently only work for one group analysis.
}
\examples{
library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
'
m1 ~ iv1 + iv2
dv ~ m1
'
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)

md_exo <- mahalanobis_exo(fit)
md_exo

}
\references{
Mahalanobis, P. C. (1936). On the generaized distance in statistics.
\emph{Proceedings of the National Institute of Science of India, 2}, 49–55.
}
