% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_rerun.R
\name{mahalanobis_rerun}
\alias{mahalanobis_rerun}
\title{Compute the Mahalanobis distance for each case}
\usage{
mahalanobis_rerun(rerun_out)
}
\arguments{
\item{rerun_out}{The output from \code{\link[=lavaan_rerun]{lavaan_rerun()}}.}
}
\value{
A one column matrix of the Mahalanobis distance for each case. The row names
are the case identification values used in \code{\link[=lavaan_rerun]{lavaan_rerun()}}.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and compute the Mahalanobis distance for each case.
}
\details{
For each case, compute the Mahalanobis distance of each case.

Currently only work for one group analysis.
}
\examples{
# To be prepared.

}
\references{
Mahalanobis, P. C. (1936). On the generaized distance in statistics. \emph{Proceedings of the National Institute of Science of India, 2}, 49–55.
}
