% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pa_dat2}
\alias{pa_dat2}
\title{Sample data set for a path model with an influential case}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\usage{
pa_dat2
}
\description{
A four-variable dataset with 100 cases and potential influential cases
}
\details{
A Sample Model:

\if{html}{\out{<div class="sourceCode">}}\preformatted{`m1 ~  iv1 + iv2`
`dv ~  m1`
}\if{html}{\out{</div>}}

Has one or more influential case
}
\keyword{datasets}
