% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pa_rerun}
\alias{pa_rerun}
\title{The \emph{n} \code{\link[lavaan:sem]{lavaan::sem()}} results on \link{pa_dat}}
\format{
An object of class \code{list} of length 3.
}
\usage{
pa_rerun
}
\description{
The \emph{n} \code{\link[lavaan:sem]{lavaan::sem()}} results on \link{pa_dat}.
}
\details{
The \code{\link[=lavaan_rerun]{lavaan_rerun()}} output fitting the following model to \link{pa_dat}\preformatted{`iv1 ~~ iv2`
`m1 ~ iv1 + iv2`
`dv ~ m1`
}
}
\keyword{datasets}
