% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lavaan_rerun.R
\name{print.lavaan_rerun}
\alias{print.lavaan_rerun}
\title{Print the results of lavaan_rerun}
\usage{
\method{print}{lavaan_rerun}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link[=lavaan_rerun]{lavaan_rerun()}}.}

\item{...}{Other arguments. They will be ignored.}
}
\value{
Nothing
}
\description{
Print the results of lavaan_rerun
}
\details{
Print the results of lavaan_rerun
}
\examples{
# TODO

}
