% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sem_dat2}
\alias{sem_dat2}
\title{Sample data set for a structural equation model with an influential case}
\format{
An object of class \code{data.frame} with 100 rows and 10 columns.
}
\usage{
sem_dat2
}
\description{
A nine-variable dataset with 100 cases and potential influential cases
}
\details{
A Sample Model:

\if{html}{\out{<div class="sourceCode">}}\preformatted{`f1 =~  x1 + x2 + x3`
`f2 =~  x4 + x5 + x6`
`f3 =~  x7 + x8 + x9`
`f2 ~   f1`
`f3 ~   f2`
}\if{html}{\out{</div>}}

Has one or more influential case
}
\keyword{datasets}
