% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_dat}
\alias{cfa_dat}
\title{Sample data set for a CFA model}
\format{
An object of class \code{data.frame} with 100 rows and 6 columns.
}
\usage{
cfa_dat
}
\description{
A six-variable dataset with 100 cases.
}
\details{
A Sample Model:

\if{html}{\out{<div class="sourceCode">}}\preformatted{`f1 =~  x1 + x2 + x3`
`f2 =~  x4 + x5 + x6`
`f1 ~~ f2`
}\if{html}{\out{</div>}}
}
\keyword{datasets}
