% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_dat}
\alias{cfa_dat}
\title{Sample Data: A CFA Model}
\format{
A data frame with 100 rows
and 6 variables:
\describe{
\item{x1}{Indicator. Numeric.}
\item{x2}{Indicator. Numeric.}
\item{x3}{Indicator. Numeric.}
\item{x4}{Indicator. Numeric.}
\item{x5}{Indicator. Numeric.}
\item{x6}{Indicator. Numeric.}
}
}
\usage{
cfa_dat
}
\description{
A six-variable dataset with 100 cases.
}
\examples{
library(lavaan)
data(cfa_dat)
mod <-
"
f1 =~  x1 + x2 + x3
f2 =~  x4 + x5 + x6
"
fit <- cfa(mod, cfa_dat)
summary(fit)
}
\keyword{datasets}
