% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_dat2}
\alias{cfa_dat2}
\title{A six-variable dataset with 100 cases and potential influential cases}
\format{
An object of class \code{data.frame} with 100 rows and 7 columns.
}
\usage{
cfa_dat2
}
\description{
A six-variable dataset with 100 cases and potential influential cases
}
\details{
Fitted by this model:\preformatted{`f1 =~  x1 + x2 + x3`
`f2 =~  x4 + x5 + x6`
`f1 ~~ f2`
}

Has one or more influential case
}
\keyword{datasets}
