% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_change.R
\name{est_change}
\alias{est_change}
\title{Standardized case influence on parameter estimates}
\usage{
est_change(rerun_out, parameters = NULL)
}
\arguments{
\item{rerun_out}{The output from \code{\link[=lavaan_rerun]{lavaan_rerun()}}.}

\item{parameters}{A character vector to specify the selected parameters.
Each parameter is named as in \code{lavaan} syntax, e.g.,
\code{x ~ y} or \code{x ~~ y}, as appeared in the columns
\code{lhs}, \code{op}, and \code{rhs} in the output of
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the default,
differences on all free parameters will be computed.}
}
\value{
A matrix with the number of columns equal to the number of requested
parameters,
and the number of rows equal to the number of cases. The row names are the
case identification values used in \code{\link[=lavaan_rerun]{lavaan_rerun()}}. The elements are the
standardized difference. Please see Pek and MacCallum (2011), Equation 7.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and compute the standardized changes in
selected parameters for each case.
}
\details{
For each case, compute the differences in the estimates of selected
parameters with and without this case: (estimate with all case) -
(estimate without this case).
The differences are standardized by dividing the raw differences by their
standard errors (Pek & MacCallum, 2011)/

If the analysis is not admissible or did not converge when a case was
deleted, \code{NA}s will be turned for this case on the differnces.

Unlike \code{\link[=est_change_raw]{est_change_raw()}}, \code{\link[=est_change]{est_change()}} does not support
computing the standardized changes of standardized estimates.

Currently only support single-sample models.
}
\examples{
library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)
# Compute the changes in chisq if a case is removed
out <- est_change(fit_rerun)
# Results excluding a case, for the first few cases
head(out)
# Note that these are the differences divided by the standard errors
# The rightmost column, `gcd`, contains the
# generalized Cook's distances (Pek & MacCallum, 2011).
head(out[, "gcd", drop = FALSE])

# Compute the changes for the paths from iv1 and iv2 to m1
out2 <- est_change(fit_rerun, c("m1 ~ iv1", "m1 ~ iv2"))
# Results excluding a case, for the first few cases
head(out2)
# Note that only the changes in the selected parameters are included.
# The generalized Cook's distance is computed only from the selected
# parameter estimates.

}
\references{
Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation
models: Cases and their influence. \emph{Multivariate Behavioral Research,
46}(2), 202–228. \url{https://doi.org/10.1080/00273171.2011.561068}

Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation
models: Cases and their influence. \emph{Multivariate Behavioral Research,
46}(2), 202–228. \url{https://doi.org/10.1080/00273171.2011.561068}
}
