% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_change_plot.R
\name{est_change_plot}
\alias{est_change_plot}
\alias{est_change_gcd_plot}
\title{Plots of Case Influence on
Parameter Estimates}
\usage{
est_change_plot(change, parameters, cutoff_change = NULL, largest_change = 1)

est_change_gcd_plot(
  change,
  parameters,
  cutoff_gcd = NULL,
  largest_gcd = 1,
  circle_size = 10
)
}
\arguments{
\item{change}{The output from
\code{\link[=est_change]{est_change()}}, \code{\link[=est_change_raw]{est_change_raw()}},
\code{\link[=est_change_approx]{est_change_approx()}}, or
\code{\link[=est_change_raw_approx]{est_change_raw_approx()}}.}

\item{parameters}{If it is
a character vector, it
specifies the selected parameters.
Each parameter is named as in
\code{lavaan} syntax, e.g., \code{x ~ y} or \code{x ~~ y}, as appeared in the columns
\code{lhs}, \code{op}, and \code{rhs} in the output
of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}.
Supports specifying an operator to
select all parameters with this
operators: \code{~}, \verb{~~}, \verb{=~}, and \code{~1}.
This vector can contain both
parameter names and operators.
If it is a numeric vector, it
specifies the columns to be used.
If omitted or \code{NULL},
the default, changes on all
parameters in \code{change}. will be used.}

\item{cutoff_change}{Cases with
absolute changes larger than this
value will be labeled. Default is
\code{NULL}. If \code{NULL}, no cutoff line
will be drawn.}

\item{largest_change}{The number of
cases with the largest absolute
changes to be labelled. Default is
\enumerate{
\item If not an integer, it will be
rounded to the nearest integer.
}}

\item{cutoff_gcd}{Cases with
generalized Cook's distance or
approximate generalized Cook's
distance larger than this value will
be labeled. Default is \code{NULL}. If
\code{NULL}, no cutoff line will be drawn.}

\item{largest_gcd}{The number of
cases with the largest generalized
Cook's distance or approximate
generalized Cook's distance to be
labelled. Default is 1. If not an
integer, it will be rounded to the
nearest integer.}

\item{circle_size}{The size of the
largest circle when the size of a
circle is controlled by a statistic.}
}
\value{
A \link{ggplot2} plot. Plotted by
default. If assigned to a variable or
called inside a function, it will not
be plotted. Use \code{\link[=plot]{plot()}} to plot it.
}
\description{
Gets the output of
functions such as \code{\link[=est_change]{est_change()}} and
\code{\link[=est_change_approx]{est_change_approx()}} and plots case
influence on selected parameters.
}
\details{
The output of
\code{\link[=est_change]{est_change()}}, \code{\link[=est_change_raw]{est_change_raw()}},
\code{\link[=est_change_approx]{est_change_approx()}}, and
\code{\link[=est_change_raw_approx]{est_change_raw_approx()}} is simply a
matrix. Therefore, these functions
will work for any matrix provided.
Row number will be used on the x-axis
if applicable. However, case
identification values will be used
for labeling individual cases if they
are stored as row names.
}
\section{Functions}{
\itemize{
\item \code{est_change_plot()}: Index plot of case influence on parameters.

\item \code{est_change_gcd_plot()}: Plot case influence against
generalized Cook's distance.

}}
\examples{

library(lavaan)

# A path model

dat <- pa_dat
mod <-
"
m1 ~ a1 * iv1 + a2 * iv2
dv ~ b * m1
a1b := a1 * b
a2b := a2 * b
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)

# Compute approximate case influence on parameters estimates
out <- est_change_approx(fit)

# Plot case influence against approximated gCD for all
# regression coefficients
# Label top 5 cases with largest approximated gCD
est_change_gcd_plot(out,
                    parameters = "~",
                    largest_gcd = 5)


}
\references{
Pek, J., & MacCallum, R.
(2011). Sensitivity analysis in
structural equation models: Cases
and their influence. \emph{Multivariate
Behavioral Research, 46}(2),
202-228.
doi:10.1080/00273171.2011.561068
}
\seealso{
\code{\link[=est_change]{est_change()}}, \code{\link[=est_change_raw]{est_change_raw()}},
\code{\link[=est_change_approx]{est_change_approx()}}, and \code{\link[=est_change_raw_approx]{est_change_raw_approx()}}.
}
\author{
Shu Fai Cheung
\url{https://orcid.org/0000-0002-9871-9448}.
}
