% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_change_raw.R
\name{est_change_raw}
\alias{est_change_raw}
\title{Case influence on parameter estimates}
\usage{
est_change_raw(rerun_out, parameters = NULL, standardized = FALSE)
}
\arguments{
\item{rerun_out}{The output from \code{\link[=lavaan_rerun]{lavaan_rerun()}}.}

\item{parameters}{A vector of characters to specify the selected parameters.
Each element is of this form: \code{x ~ y} or \code{x ~~ y},
corresponds to how each parameter is specified in \code{lavaan}.
The naming convention of a \code{lavaan} output can be found
by \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the default,
differences on all parameters will be computed.}

\item{standardized}{If \code{TRUE}, the changes in full standardized is returned.
Otherwise, the changes in unstandardized solution is returned.
Default is \code{FALSE}.}
}
\value{
A matrix with the number of columns equal to the number of requested parameters,
and the number of rows equal to the number of cases. The row names is the
case identification values used in \code{\link[=lavaan_rerun]{lavaan_rerun()}}. The elements are the
standardized difference. Please see Pek and MacCallum (2011), Equation 7.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and compute the changes in selected parameters
for each case.
}
\details{
For each case, compute the differences in the estimates of selected parameter
between rerun without
this case and the original fit with this case.
The change is the raw change, either for the standardized or
unstandardized solution. The change is \emph{not} divided by standard error.#'
Currently only work for one group analysis.
}
\examples{
# To be prepared.

}
\references{
Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. \emph{Multivariate Behavioral Research, 46}(2), 202–228. \url{https://doi.org/10.1080/00273171.2011.561068}
}
