% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_change_raw_approx.R
\name{est_change_raw_approx}
\alias{est_change_raw_approx}
\title{Case Influence on Parameter Estimates (Approximated)}
\usage{
est_change_raw_approx(fit, parameters = NULL, case_id = NULL)
}
\arguments{
\item{fit}{The output from \code{\link[lavaan:lavaan]{lavaan::lavaan()}}.}

\item{parameters}{A character vector to specify the selected
parameters. Each parameter is named as in \code{lavaan} syntax, e.g.,
\code{x ~ y} or \code{x ~~ y}, as appeared in the columns \code{lhs}, \code{op}, and \code{rhs}
in the output of \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}.
Supports specifying an operator to select all parameters with this
operators: \code{~}, \verb{~~}, \verb{=~}, and \code{~1}. This vector can contain
both parameter names and operators.
If \code{NULL}, the
default, differences on all free parameters will be computed.}

\item{case_id}{If it is a character vector of length equals to the
number of cases (the number of rows in the data in \code{fit}), then it
is the vector of case identification values. If it is \code{NULL}, the
default, then \code{case.idx} used by \code{lavaan} functions will be used
as case identification values.}
}
\value{
A matrix with the number of columns equals to the number of
requested parameters, and the number of rows equals to the number
of cases. The row names are case identification values. The
elements are the raw differences.
}
\description{
Gets a \code{\link[lavaan:lavaan]{lavaan::lavaan()}} output and estimates the
changes in selected parameters for each case.
}
\details{
For each case, \code{\link[=est_change_raw_approx]{est_change_raw_approx()}} computes the
approximated differences
in the estimates of selected parameters with and without this
case: (estimate with all case) - (estimate without this case). The
change is the raw change, either for the standardized or
unstandardized solution. The change is \emph{not} divided by standard
error.

The model is not refitted. Therefore, the result is only an
approximation of that of \code{\link[=est_change_raw]{est_change_raw()}}. However, this
approximation is useful for identifying potentially influential
cases when the sample size is very large or the model takes a long
time to fit. This function can be used to identify potentially
influential cases quickly and then select them to conduct the
leave-one-out sensitivity analysis using \code{\link[=lavaan_rerun]{lavaan_rerun()}} and
\code{\link[=est_change_raw]{est_change_raw()}}.

Currently it only supports single-group models.
}
\examples{
library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
# Compute the approximated changes in parameter estimates if a case is removed
out_approx <- est_change_raw_approx(fit)
head(out_approx)
# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)
# Compute the changes in parameter estimates if a case is removed
out <- est_change_raw(fit_rerun)
# Results excluding a case, for the first few cases
head(out)
# Compare the results
plot(out_approx[, 1], out[, 1]); abline(a = 0, b = 1)
plot(out_approx[, 5], out[, 5]); abline(a = 0, b = 1)

}
\author{
Idea by Mark Hok Chio Lai \url{https://orcid.org/0000-0002-9196-7406},
Implemented by Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}.
}
