% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_measures_change.R
\name{fit_measures_change}
\alias{fit_measures_change}
\title{Case influence on fit measures from \code{lavaan} fitMeasures}
\usage{
fit_measures_change(
  rerun_out,
  fit_measures = c("chisq", "cfi", "tli"),
  baseline_model = NULL
)
}
\arguments{
\item{rerun_out}{The output from \code{\link[=lavaan_rerun]{lavaan_rerun()}}.}

\item{fit_measures}{The argument \code{fit.measures} used in
\link[lavaan:fitMeasures]{lavaan::fitMeasures}. Default
is \code{c("chisq", "cfi", "tli")}.}

\item{baseline_model}{The argument \code{baseline.model}     used in
\link[lavaan:fitMeasures]{lavaan::fitMeasures}. Default}
}
\value{
A matrix with the number of columns equal to the number of requested fit
measures,
and the number of rows equal to the number of cases. The row names is the
case identification values used in \code{\link[=lavaan_rerun]{lavaan_rerun()}}.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and compute the change in selected fit
measures
for each case.
}
\details{
For each case, compute the difference in selected fit measures between
rerun without
this case and the original fit with this case.

If the analysis is not admissible or did not converge when a case was
deleted, \code{NA} will be turned for this case.

Currently only work for one group analysis.
}
\examples{
library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <- 
'
m1 ~ iv1 + iv2
dv ~ m1
'
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)
# Compute the changes in chisq if a case is removed
out <- fit_measures_change(fit_rerun, fit_measures = "chisq")
# Results excluding a case, for the first few cases
head(out)
# Chi-square will all cases
(chisq_all <- fitMeasures(fit, c("chisq")))
# Chi-square with the first case removed
fit_01 <- lavaan::sem(mod, dat[-1, ])
(chisq_no_1 <- fitMeasures(fit_01, c("chisq")))
# Difference
chisq_all - chisq_no_1
# Compare the result from the function
out[1, ]
}
\references{
Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation
models: Cases and their influence. \emph{Multivariate Behavioral Research,
46}(2), 202–228. \url{https://doi.org/10.1080/00273171.2011.561068}
}
