% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_scores.R
\name{implied_scores}
\alias{implied_scores}
\title{Compute the implied scores for an SEM model}
\usage{
implied_scores(fit)
}
\arguments{
\item{fit}{The output from \code{lavaan}, such as \code{\link[lavaan:cfa]{lavaan::cfa()}} and
\code{\link[lavaan:sem]{lavaan::sem()}}.}
}
\value{
A matrix of the implied scores
}
\description{
Get a \code{lavaan} output and compute the implied scores.
}
\details{
It currently supports single-group path analytic models only.
}
\examples{
library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)

# (To be continued)


}
