% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_stat.R
\name{influence_stat}
\alias{influence_stat}
\title{Compute case influence}
\usage{
influence_stat(
  rerun_out,
  fit_measures = c("chisq", "cfi", "rmsea", "tli"),
  baseline_model = NULL,
  parameters = NULL,
  mahalanobis = TRUE,
  keep_fit = TRUE
)
}
\arguments{
\item{rerun_out}{The output from \code{\link[=lavaan_rerun]{lavaan_rerun()}}.}

\item{fit_measures}{The argument \code{fit.measures} used in
\link[lavaan:fitMeasures]{lavaan::fitMeasures}. Default
is \code{c("chisq", "cfi", "rmsea", "tli")}.}

\item{baseline_model}{The argument \code{baseline.model}     used in
\link[lavaan:fitMeasures]{lavaan::fitMeasures}. Default}

\item{parameters}{A character vector to specify the selected parameters.
Each parameter is named as in \code{lavaan} syntax, e.g.,
\code{x ~ y} or \code{x ~~ y}, as appeared in the columns
\code{lhs}, \code{op}, and \code{rhs} in the output of
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the default,
differences on all free parameters will be computed.}

\item{mahalanobis}{If \code{TRUE}, will call \code{\link[=mahalanobis_rerun]{mahalanobis_rerun()}} to compute the
Mahalanobis distance. Default is \code{TRUE}.}

\item{keep_fit}{If \code{TRUE}, will keep the original \code{lavaan} output using the
full
sample as an attribute to the output. It can be used by
other functions to
extract necessary information. Default is \code{TRUE}.}
}
\value{
A matrix with the number of columns equal to the number of requested
statistics,
and the number of rows equal to the number of cases. The row names is the
case identification values used in \code{\link[=lavaan_rerun]{lavaan_rerun()}}. Please refer to the
corresponding
functions for further details.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and compute the changes in selected parameters
and fit measures for each case.
}
\details{
For each case, compute the differences in the estimates of selected
parameters
and fit measures
with and without this case. Can also request measures of
extremeness (only Mahalanobis distance is available for now).

It is a wrapper of \code{\link[=est_change]{est_change()}} and \code{\link[=fit_measures_change]{fit_measures_change()}}.

Currently only work for single-sample models.
}
\examples{
library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)
# Get all default influence stats
out <- influence_stat(fit_rerun)
head(out)

}
\references{
Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation
models: Cases and their influence. \emph{Multivariate Behavioral Research,
46}(2), 202–228. \url{https://doi.org/10.1080/00273171.2011.561068}
}
\seealso{
\code{\link[=fit_measures_change]{fit_measures_change()}}, \code{\link[=est_change]{est_change()}}, and \code{\link[=mahalanobis_rerun]{mahalanobis_rerun()}}.
}
