% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_stat.R
\name{influence_stat}
\alias{influence_stat}
\title{Compute case influence on parameter estimates and fit measures and extremeness}
\usage{
influence_stat(
  rerun_out,
  fit_measures = c("chisq", "cfi", "tli"),
  baseline_model = NULL,
  parameters = NULL,
  mahalanobis = TRUE,
  keep_fit = TRUE
)
}
\arguments{
\item{rerun_out}{The output from \code{\link[=lavaan_rerun]{lavaan_rerun()}}.}

\item{fit_measures}{The argument \code{fit.measures} used in
\link[lavaan:fitMeasures]{lavaan::fitMeasures}. Default
is \code{c("chisq", "cfi", "tli")}. If \code{FALSE}, changes
in fit measures will not be computed.}

\item{baseline_model}{The argument \code{baseline.model}     used in
\link[lavaan:fitMeasures]{lavaan::fitMeasures}. Default}

\item{parameters}{A vector of characters to specify the selected parameters.
Each element is of this form: \code{x ~ y} or \code{x ~~ y},
corresponds to how each parameter is specified in \code{lavaan}.
The naming convention of a \code{lavaan} output can be found
by \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. If \code{NULL}, the default,
differences on all parameters will be computed.
If \code{FALSE}, changes in parameter estimates will not be computed.}

\item{mahalanobis}{If \code{TRUE}, will call \code{\link[=mahalanobis_rerun]{mahalanobis_rerun()}} to compute the
Mahalanobis distance. Default is \code{TRUE}.}

\item{keep_fit}{If \code{TRUE}, will keep the original \code{lavaan} output will full
sample as an attribute to the output. For other functions to
extract necessary information. Default is \code{TRUE}.}
}
\value{
A matrix with the number of columns equal to the number of requested statistics,
and the number of rows equal to the number of cases. The row names is the
case identification values used in \code{\link[=lavaan_rerun]{lavaan_rerun()}}. Please refer to the corresponding
functions for further details.
}
\description{
Get a \code{\link[=lavaan_rerun]{lavaan_rerun()}} output and compute the changes in selected parameters
and fit measures for each case.
}
\details{
For each case, compute the differences in the estimates of selected parameters
and fit measures
between rerun without
this case and the original fit with this case. Can also request measures of
extremeness (only Mahalanobis distance is available for now).

A wrapper of \code{\link[=est_change]{est_change()}} and \code{\link[=fit_measures_change]{fit_measures_change()}}.

Currently only work for one group analysis.
}
\examples{
# To be prepared.

}
\references{
Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. \emph{Multivariate Behavioral Research, 46}(2), 202–228. \url{https://doi.org/10.1080/00273171.2011.561068}
}
\seealso{
\code{\link[=fit_measures_change]{fit_measures_change()}}, \code{\link[=est_change]{est_change()}}, and \code{\link[=mahalanobis_rerun]{mahalanobis_rerun()}}.
}
