% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pa_dat}
\alias{pa_dat}
\title{A four-variable dataset with 100 cases}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
pa_dat
}
\description{
A four-variable dataset with 100 cases
}
\details{
A Sample Model:\preformatted{`m1 ~  iv1 + iv2`
`dv ~  m1`
}
}
\keyword{datasets}
