% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars_id.R
\name{pars_id}
\alias{pars_id}
\title{Convert Parameter Syntax to Ids
or Row Numbers in the Parameter
Vector or Table}
\usage{
pars_id(pars, fit, where = c("coef", "partable"), free_only = TRUE)
}
\arguments{
\item{pars}{A character vector of parameters specified
in lavaan syntax, e.g., \code{"y ~ x"} and \code{f1 =~ x3}. For
multisample models, if only the parameters in some groups
are needed, use the modifier for labeling parameters and
use \code{NA} to denote parameters to be requested. E.g.,
\code{f1 =~ c(NA, 0, NA, NA) * x2} denotes the loadings of
\code{x2} on \code{f1} in the first, third, and fourth groups.}

\item{fit}{A \code{lavaan}-class object. This object is used
to determine the number of groups and the parameters
in the model. Only parameters in \code{pars} that appear in
this model will be considered.}

\item{where}{Where the values are to be found. Can be
"partable" (parameter table) or "coef"
(coefficient vector).
Default is "coef".}

\item{free_only}{Wether only free parameters will be
kept. Default is \code{TRUE}.}

\item{type}{Whether the ids are based on the positions
in the vector of
free parameters ("free"), as in the output of \code{coef()},
or all parameters ("all") in the parameter table. Used
only for parameters specified by operators and \code{where}
is "partable".
Default is "free".}
}
\value{
A numeric vector of the ids in the column "free" in the
parameter table of the fit object.
}
\description{
An internal function to convert a vector of
parameters specified in lavaan syntax to the ids of them
in the vector of free parameters or the row numbers
in the parameter table.
}
\details{
It uses \code{\link[lavaan:model.syntax]{lavaan::lavaanify()}} to parse the
syntax strings. Multiple sample models is supported.
}
\examples{
\donttest{
# To Prepare.
}

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
