% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars_id.R
\name{pars_id}
\alias{pars_id}
\title{Convert Parameter Syntax to Position
or Row Numbers in the Parameter
Vector or Table}
\usage{
pars_id(pars, fit, where = c("coef", "partable"), free_only = TRUE)
}
\arguments{
\item{pars}{A character vector of parameters specified
in lavaan syntax, e.g., \code{"y ~ x"} and \code{f1 =~ x3}. For
multisample models, if only the parameters in some groups
are needed, use the modifier for labeling parameters and
use \code{NA} to denote parameters to be requested. E.g.,
\code{f1 =~ c(NA, 0, NA, NA) * x2} denotes the loadings of
\code{x2} on \code{f1} in the first, third, and fourth groups.}

\item{fit}{A \code{lavaan}-class object. This object is used
to determine the number of groups and the parameters
in the model. Only parameters in \code{pars} that appear in
this model will be considered.}

\item{where}{Where the values are to be found. Can be
"partable" (parameter table) or "coef"
(coefficient vector).
Default is "coef".}

\item{free_only}{Wether only free parameters will be
kept. Default is \code{TRUE}.}
}
\value{
A numeric vector of the ids in the column "free" in the
parameter table of the fit object.
}
\description{
It converts a vector of
parameters specified in lavaan syntax to the ids of them
in the vector of free parameters or the row numbers
in the parameter table.
}
\details{
It supports the following ways to specify
the parameters to be included.
}
\section{\code{lavaan} syntax}{
For example, \code{"y ~ x"} denotes the regression coefficient
regressin \code{y} on \code{x}. It uses \code{\link[lavaan:model.syntax]{lavaan::lavaanify()}} to
parse the syntax strings.
}

\section{Operator}{
For example, \code{"~"} denotes all regression coefficients.
}

\section{Multisample model}{
If a model has more than one groups, a specification
specified as in a single sample model denotes the same
parameters in all group.

For example, \code{"f1 =~ x2"} denote the factor loading of
\code{x2} on \code{f1} in all groups. \code{"~~"} denotes covariances
and error covariances in all groups.

There are two ways to select parameters only in selected
groups. First, the syntax to fix parameter values
can be used, with \code{NA} denoting parameters to be selected.

For example, \code{"f2 =~ c(NA, 1, NA) * x5"} select the
factor loadings of \code{x5} on \code{f2} in the first and third
groups.

Users can also add ".grouplabel" to a specification,
\code{grouplabel} being the group label of a group (the one
appears in \code{\link[=summary]{summary()}}, not the one of the form \code{".gp2"},
\code{"gp3"}, etc.).

For example, \code{"f2 =~ x5.Alpha"} denotes the factor loading
of \code{x5} on \code{f2} in the group \code{"Alpha"}.
}

\examples{

dat <- sem_dat
set.seed(64264)
dat$gp <- sample(c("Alpha", "Beta", "Gamma"),
                 nrow(dat),
                 replace = TRUE)

library(lavaan)
sem_model <-
"
f1 =~  x1 + x2 + x3
f2 =~  x4 + x5 + x6
f3 =~  x7 + x8 + x9
f2 ~   f1
f3 ~   f2
"

fit_ng <- sem(sem_model, dat)
fit_gp <- sem(sem_model, dat, group = "gp")

pars <- c("f1 =~ x2", "f2 =~ x5", "f2 ~ f1")
tmp <- pars_id(pars, fit = fit_ng)
coef(fit_ng)[tmp]
tmp <- pars_id(pars, fit = fit_ng, where = "partable")
parameterTable(fit_ng)[tmp, ]

pars <- c("f1 =~ x2", "f2 =~ c(NA, 1, NA) * x5")
tmp <- pars_id(pars, fit = fit_gp)
coef(fit_gp)[tmp]
tmp <- pars_id(pars, fit = fit_gp, where = "partable")
parameterTable(fit_gp)[tmp, ]

pars2 <- c("f1 =~ x2", "~~.Beta", "f2 =~ x5.Gamma")
tmp <- pars_id(pars2, fit = fit_gp)
coef(fit_gp)[tmp]
tmp <- pars_id(pars2, fit = fit_gp, where = "partable")
parameterTable(fit_gp)[tmp, ]
# Note that group 1 is "Beta", not "Alpha"
lavInspect(fit_gp, "group.label")

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
