% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars_id.R
\name{pars_id_to_lorg}
\alias{pars_id_to_lorg}
\title{Convert Ids to "lhs-op-rhs-(group)"}
\usage{
pars_id_to_lorg(pars_id, pars_source, type = c("free", "all"))
}
\arguments{
\item{pars_id}{A vector of integers. Usually the output
of \link{pars_id}.}

\item{pars_source}{Can be the output of
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}} or \code{\link[lavaan:lavParTable]{lavaan::parameterTable()}},
or a named vector of free parameters (e.g., the output
of \code{\link[=coef]{coef()}} applied to a \code{lavaan}-class object).}

\item{type}{The meaning of the values in \code{pars_id}.
If "free", they
are the position in the vector of free parameters (i.e.,
the output of of \code{coef()}). If "all", they are the
row numbers in the parameter table (the output of
\code{\link[lavaan:lavParTable]{lavaan::parameterTable()}}).}
}
\description{
Convert ids generated by \code{\link[=pars_id]{pars_id()}}
to values that can be used to extract the parameters
from the source.
}
\details{
If the source is a parameter estimates table, it returns
a data frame with columns "lhs", "op", and "rhs". If
"group" is present in the source, it also add a column
"group". These columns can be used to uniquely identify
the parameters specified by the ids.

If the source is a named vector of parameters (e.g., the
output of \code{\link[=coef]{coef()}}), it returns the names of parameters
based on the ids.
}
\examples{
\donttest{
# To Prepare.
}
}
