% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sem_dat}
\alias{sem_dat}
\title{A nine-variable dataset with 100 cases}
\format{
An object of class \code{data.frame} with 200 rows and 9 columns.
}
\usage{
sem_dat
}
\description{
A nine-variable dataset with 100 cases.
}
\details{
A Sample Model:\preformatted{`f1 =~  x1 + x2 + x3`
`f2 =~  x4 + x5 + x6`
`f3 =~  x7 + x8 + x9`
`f2 ~   f1`
`f3 ~   f2`
}
}
\keyword{datasets}
