% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sem_dat}
\alias{sem_dat}
\title{A four-variable dataset with 100 cases}
\format{
An object of class \code{data.frame} with 200 rows and 9 columns.
}
\usage{
sem_dat
}
\description{
A four-variable dataset with 100 cases.
}
\details{
Generated from this model:\preformatted{`f1 =~  .7*x1 + .6*x2 + .8*x3 + .3*x5`
`f2 =~  .2*x1 + .6*x4 + .8*x5 + .7*x6`
`f3 =~  .5*x4 + .2*x7 + .6*x8 + .8*x9`
`f2 ~   .3*f1`
`f3 ~   .3*f2 + .8*f1`
}
}
\keyword{datasets}
