% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semfindr-package.R
\docType{package}
\name{semfindr-package}
\alias{semfindr}
\alias{semfindr-package}
\title{semfindr: Compute diagnostic statistics for outliers and influential cases in structural equation modeling}
\description{
Compute diagnostic statistics for outliers and influential cases in structural equation modeling.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/sfcheung/semfindr/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

}
\keyword{internal}
