% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semfindr-package.R
\docType{package}
\name{semfindr-package}
\alias{semfindr}
\alias{semfindr-package}
\title{semfindr: Diagnostic Tools in Structural Equation Modeling}
\description{
Sensitivity analysis in structural equation modeling using change statistics and diagnostic plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/semfindr/}
  \item Report bugs at \url{https://github.com/sfcheung/semfindr/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Mark Hok Chio Lai \email{marklhc@gmail.com } (\href{https://orcid.org/0000-0002-9196-7406}{ORCID})
}

}
\keyword{internal}
