% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semfindr-package.R
\docType{package}
\name{semfindr-package}
\alias{semfindr}
\alias{semfindr-package}
\title{semfindr: Influential Cases in Structural Equation Modeling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sensitivity analysis in structural equation modeling using influence measures and diagnostic plots. Support leave-one-out casewise sensitivity analysis presented by Pek and MacCallum (2011) \doi{10.1080/00273171.2011.561068} and approximate casewise influence using scores and casewise likelihood.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/semfindr/}
  \item Report bugs at \url{https://github.com/sfcheung/semfindr/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Mark Hok Chio Lai \email{marklhc@gmail.com } (\href{https://orcid.org/0000-0002-9196-7406}{ORCID})
}

}
\keyword{internal}
